pyproject.toml
[tool.poetry]
name = "ai-demo"
version = "0.1.0"
description = ""
authors = ["Samuel Davis <sam99dave@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-fasthtml = "^0.6.13"
ipykernel = "^6.29.5"
requests = "^2.32.3"
pillow = "^11.0.0"
supabase = "^2.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
